Add loadit tool for testing sensu installations
