+++ cmd/loadit/README.md
@@ -1,0 +1,1 @@
+The loadit tool is for load testing sensu installations.
